requirements.txt
pillow==10.4.0
PyPDF2==2.12.0
transformers==4.43.2
regex==2024.5.15